Further abstract publisher ABC
