+++ tests/helpers/publisher/mqtt/__init__.py
@@ -1,0 +1,1 @@
+"""Define tests for MQTT publishers."""
